Move messaging actions into their own folder
