+++ packages/wallet/src/redux/protocols/dispute/responder/__tests__/reducer.test.ts
@@ -6,7 +6,7 @@ import * as TransactionGenerator from "../../../../../utils/transaction-generato
 import {itSendsThisDisplayEventType, itSendsThisMessage} from "../../../../__tests__/helpers";
 import {describeScenarioStep} from "../../../../__tests__/helpers";
 import {State} from "@statechannels/nitro-protocol";
-import {apiNotImplemented} from "../../../../actions";
+import {apiNotImplemented} from "../../../../sagas/messaging/outgoing-api-actions";
 
 // Mocks
 const mockTransaction = {to: "0xabc"};
