+++ packages/wallet/src/redux/sagas/messaging/message-sender.ts
@@ -1,13 +1,13 @@
-import {OutgoingApiAction, messageSent} from "../actions";
-
+import {messageSent} from "../../actions";
+import {OutgoingApiAction} from "./outgoing-api-actions";
 import {call, select, put} from "redux-saga/effects";
-import {getChannelStatus} from "../state";
-import {ChannelState, getLastState} from "../channel-store";
-import {createJsonRpcAllocationsFromOutcome} from "../../utils/json-rpc-utils";
+import {getChannelStatus} from "../../state";
+import {ChannelState, getLastState} from "../../channel-store";
+import {createJsonRpcAllocationsFromOutcome} from "../../../utils/json-rpc-utils";
 import jrs from "jsonrpc-lite";
-import {unreachable} from "../../utils/reducer-utils";
-import {validateResponse, validateNotification} from "../../json-rpc-validation/validator";
-import {getChannelHoldings} from "../selectors";
+import {unreachable} from "../../../utils/reducer-utils";
+import {validateResponse, validateNotification} from "../../../json-rpc-validation/validator";
+import {getChannelHoldings} from "../../selectors";
 import {bigNumberify} from "ethers/utils";
 
 export function* messageSender(action: OutgoingApiAction) {
